inclusão das notações para o banco de dados

+++ RestWithASPNETUdemy/Service/IPersonService.cs
@@ -4,8 +4,8 @@ namespace RestWithASPNETUdemy.Services;
 
 public interface IPersonService{
     Person Create(Person person);
-    Person FindByID(long id);
+    Person FindByID(int id);
     List<Person> FindAll();
     Person Update(Person person);
-    void Delete(long id);
+    void Delete(int id);
 }
